Final EDA Report
I finished the report. The appendix is not the prettiest. If you want, you can go through and make edits. and also just make sure we have all the boxes checked for what she wants. if you could put a couple sentences in (maybe in the appendix) about how you went about merging things, thatd be good. also theres a (1) AND (2) for Part 2. (2) is the report. (1) is cleaning and graphical/numerical summaries. I was just gonna turn in the EDA R markdown I used and the merge R file